// models/profile/formdata-profile-cv-model.ts
export class FormDataProfileCVModel { 
    profilePic: string = "";
    statusVerify: string = "";
    aboutMe: string = "";
    email: string = "";
    phone: string = "";
    skill: string = "";
    expectedSalary: string = "";
    lastEdu: string = "";
    pkPosition: string[] = []
    pkYear: string[] = []
    pkCompany: string[] = []
}